+++ src/game/items.h
@@ -13,5 +13,6 @@ void __cdecl Item_NewRoom(int16_t item_num, int16_t room_num);
 int32_t __cdecl Item_GlobalReplace(
     int32_t src_object_num, int32_t dst_object_num);
 void __cdecl Item_ClearKilled(void);
+bool Item_Teleport(ITEM_INFO *item, int32_t x, int32_t y, int32_t z);
 
 bool Item_IsSmashable(const ITEM_INFO *item);