+++ src/game/room.c
@@ -1,0 +1,21 @@
+#include "game/room.h"
+
+#include "global/const.h"
+#include "global/vars.h"
+
+int16_t Room_GetIndexFromPos(const int32_t x, const int32_t y, const int32_t z)
+{
+    for (int i = 0; i < g_RoomCount; i++) {
+        const ROOM_INFO *const room = &g_Rooms[i];
+        const int32_t x1 = room->pos.x + WALL_L;
+        const int32_t x2 = room->pos.x + (room->y_size << WALL_SHIFT) - WALL_L;
+        const int32_t y1 = room->max_ceiling;
+        const int32_t y2 = room->min_floor;
+        const int32_t z1 = room->pos.z + WALL_L;
+        const int32_t z2 = room->pos.z + (room->x_size << WALL_SHIFT) - WALL_L;
+        if (x >= x1 && x < x2 && y >= y1 && y <= y2 && z >= z1 && z < z2) {
+            return i;
+        }
+    }
+    return NO_ROOM;
+}