+++ src/game/room.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+
+int16_t Room_GetIndexFromPos(int32_t x, int32_t y, int32_t z);