+++ src/game/shell.c
@@ -1,6 +1,7 @@
 #include "game/shell.h"
 
 #include "decomp/decomp.h"
+#include "game/console.h"
 #include "game/input.h"
 #include "game/music.h"
 #include "global/funcs.h"