+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased](https://github.com/LostArtefacts/TR2X/compare/stable...develop) - ××××-××-××
+- added dev console with the following commands:
+    - `/pos`
+    - `/tp [room_num]`
+    - `/tp [x] [y] [z]`
 
 ## [0.1.1](https://github.com/LostArtefacts/TR2X/compare/0.1...0.1.1) - 2024-04-27
 - fixed Lara's shadow with z-buffer option on (#64, regression from 0.1)